feat: display articulation items in rows
