+++ next/src/components/articulation/overview/layout/ArticulationOverviewVersionLayout.tsx
